Refactoring some common structures and functions
- Created common.h to hold common functions between files, like
printPath, movePlayer
- Moved some common typedefs to structures.h
- Removed movesBlocked from sokoban_heuristic to refactor boxTrapped
function
- Created common addPath, boxTrapped and checkWallsAt on common.c,
without exposing to algorithms
- Fixed sokoban_heuristic.c to find solutions. Works only with level_-1
right now.
