+++ brute-force/single-thread/Sokoban_HashTableList.c
@@ -1,4 +1,6 @@
 #include "../../common/sort.h"
+#include "../../common/structures.h"
+#include "../../common/common.h"
 #include "../../common/util.h"
 #include <stdio.h>
 #include <stdlib.h>
@@ -7,27 +9,6 @@
 
 // Valor do módulo usado no hashing do id (que por si só é um hash)
 #define ID_HASH 100000000
-
-// Tamanho máximo do caminho
-#define MAX_PATH_SIZE 3000
-
-// Estado que estaremos analisando
-typedef struct State {
-	// Todas as posições se baseiam no fato que não há mapa com mais de 19 de
-	// largura/altura. Portanto, a posição é 20*y+x;
-	unsigned short posPlayer;
-	unsigned short posBoxes[30];
-	unsigned short posGoals[30];
-	unsigned char grid[400];
-	// Quantidade de caixas encontradas
-	unsigned char boxes;
-	// Quantidade de caixas no objetivo
-	unsigned char boxesOnGoals;
-	// Caminho para a solução
-	unsigned char path[MAX_PATH_SIZE];
-	unsigned int pathSize;
-	unsigned int heuristic;
-} State;
 
 // Lista ligada (com hash de resto da divisão) para os ids visitados.
 typedef struct visitedId {
@@ -49,8 +30,6 @@ visitedId *idList[ID_HASH];
 
 // Ponteiro para o último nó
 Node **last;
-
-void printPath(State *s) { printf("%s\n", s->path); }
 
 // Função que procura o id na lista
 unsigned char findId(char *idHash, unsigned long long int id) {
@@ -146,15 +125,6 @@ unsigned char getStateId(State *s) {
 	return 0;
 };
 
-// Adicionamos o movimento feito na lista de movimentos
-void addPath(char c, State *s) {
-	// Se atingimos o limite de tamanho, encerramos
-	if (s->pathSize == MAX_PATH_SIZE) {
-		return;
-	}
-	s->path[s->pathSize++] = c;
-};
-
 //
 void placeThis(char c, int x, int y, struct State *s) {
 	if (c == 32) {
@@ -212,10 +182,8 @@ void buildMap(struct State *s, char *level) {
 	// Inicializamos algumas das variáveis de State
 	s->boxes = 0;
 	s->boxesOnGoals = 0;
-	s->pathSize = 0;
 	memset(s->posGoals, 0, 30);
 	memset(s->posBoxes, 0, 30);
-	memset(s->path, 0, 200);
 
 	// Apontamos cada um dos Ids para nulo.
 	memset(idList, 0, ID_HASH);
@@ -259,151 +227,6 @@ void buildMap(struct State *s, char *level) {
 	fclose(file);
 
 	getStateId(s);
-};
-
-// Função simples que retorna se a posição pos é uma parede
-int checkWallsAt(State *s, unsigned int pos) {
-	if (s->grid[pos] == '#') {
-		return 1;
-	}
-	return 0;
-};
-
-// Função que verifica se prendemos a caixa
-/*
-     #
-    #$
-    Procuramos "cantos" como este. Se a caixa está num canto, e este canto não é
-    um objetivo, travamos a caixa Esta é uma posição exemplo que não seria
-    explorada:
-    #######
-    #   @$# <- Note que a caixa não poderá ser movida
-    #     #
-    #. #  #
-    #. $  #
-    #.$$  #
-    #.#  @#
-    #######
-*/
-unsigned char boxTrapped(State *s, int pos) {
-	int wallsUp = s->grid[pos - 20] == '#',
-	    wallsDown = s->grid[pos + 20] == '#',
-	    wallsLeft = s->grid[pos - 1] == '#',
-	    wallsRight = s->grid[pos + 1] == '#';
-	if (s->grid[pos] == '.') {
-		return 0;
-	}
-	if (wallsUp && wallsLeft) {
-		return 1;
-	}
-	if (wallsUp && wallsRight) {
-		return 1;
-	}
-	if (wallsDown && wallsLeft) {
-		return 1;
-	}
-	if (wallsDown && wallsRight) {
-		return 1;
-	}
-
-	return 0;
-}
-
-// Função que usaremos intensivamente para mover o sokoban
-char movePlayer(struct State *s, int dir) {
-	// dir == 0 -> direita
-	// dir == 1 -> esquerda
-	// dir == 2 -> cima
-	// dir == 3 -> baixo
-	char c;
-	unsigned char box = 0;
-	int tempPos, i = -1;
-	int movingParam;
-	switch (dir) {
-	case direita:
-		movingParam = 1;
-		c = 'r';
-		break;
-	case esquerda:
-		movingParam = -1;
-		c = 'l';
-		break;
-	case cima:
-		movingParam = -20;
-		c = 'u';
-		break;
-	case baixo:
-		movingParam = 20;
-		c = 'd';
-		break;
-	}
-
-	// Usamos o trecho de código para obter o parametro de movimento, assim como
-	// o caracter. Como parâmetro de movimento, queremos transformar a visao 1D
-	// do array no movimento 2D do personagem, ou seja, dado que cada linha
-	// possui 20 colunas, "andar para cima" recua 20 colunas, por isso o
-	// movingParam é -20. A mesma lógica se aplica para os outros
-
-	tempPos = s->posPlayer + movingParam;
-
-	if (checkWallsAt(s, tempPos)) {
-		// Tem uma parede.
-		return 0;
-	}
-
-	if (s->grid[tempPos] == '$' || s->grid[tempPos] == '*') {
-		// Tem uma caixa na direção
-		if (checkWallsAt(s, tempPos + movingParam) == 1 ||
-		    s->grid[tempPos + movingParam] == '$' ||
-		    s->grid[tempPos + movingParam] == '*') {
-			// Tem uma parede ou caixa após a caixa.
-			return 0;
-		};
-		// Deixa a letra maiuscula
-		c -= 32;
-		box = 1;
-	}
-
-	// Efetiva o movimento
-	if (box != 0) {
-		if (boxTrapped(s, tempPos + movingParam)) {
-			return 0;
-		}
-		// Empurrou uma caixa
-		if (s->grid[tempPos] == '*') {
-			// A caixa estava sobre um alvo
-			s->grid[tempPos] = '.';
-			s->boxesOnGoals--;
-		} else {
-			// Não estava sobre nada
-			s->grid[tempPos] = 32;
-		}
-		if (s->grid[tempPos + movingParam] == '.') {
-			// A posição de destino da caixa é alvo
-			s->grid[tempPos + movingParam] = '*';
-			s->boxesOnGoals++;
-		} else {
-			s->grid[tempPos + movingParam] = '$';
-		}
-		// Move a caixa nas posições
-		for (int i = 0; i < 30 && s->posBoxes[i] != 0; i++) {
-			if (s->posBoxes[i] == tempPos) {
-				s->posBoxes[i] += movingParam;
-			}
-		}
-	}
-
-	s->posPlayer = tempPos;
-
-	// Verifica se o Id é único e claro, qual seu valor.
-	if (getStateId(s)) {
-		return 0;
-	}
-
-	addPath(c, s);
-
-	// Retorna o efeito
-	return c;
 };
 
 // Função que verifica se o estado é final
@@ -507,7 +330,7 @@ int main(int argc, char *argv[]) {
 			copyState(rootState, s);
 
 			// Se foi possível mover, insere o estado
-			if (movePlayer(s, i) != 0) {
+			if (movePlayer(s, i, getStateId) != 0) {
 
 				// Verifica se este estado é final, junto com inseri-lo na lista
 				final = insertState(&root, s);
