+++ common/common.h
@@ -1,0 +1,12 @@
+#include "structures.h"
+
+#ifndef _COMMON_H_
+#define _COMMON_H_
+
+// Função que imprime o caminho para a solução
+extern void printPath(State *s);
+
+// Função que usaremos intensivamente para mover o sokoban
+extern char movePlayer(State *s, int dir, unsigned char (*getStateId)(State *));
+
+#endif
