+++ common/util.h
@@ -3,6 +3,10 @@
 
 #define copyState(s, s2) memcpy(s2, s, sizeof(State))
 #define NANOS ((unsigned long)1000000000)
+
+// Tamanho máximo do caminho
+#define MAX_PATH_SIZE 3000
+
 enum Direcoes { direita, esquerda, cima, baixo };
 
 #endif
