+++ heuristic/sokoban_heuristic.c
@@ -1,4 +1,6 @@
 #include "../common/sort.h"
+#include "../common/structures.h"
+#include "../common/common.h"
 #include "../common/util.h"
 #include <stdio.h>
 #include <stdlib.h>
@@ -6,25 +8,6 @@
 
 #define MAX_Node 100000000
 #define ID_HASH 100000000
-#define MAX_PATH_SIZE 3000
-
-// Estado que estaremos analisando
-typedef struct State {
-	// Todas as posições se baseiam no fato que não há mapa com mais de 19 de
-	// largura/altura. Portanto, a posição é 20*y+x;
-	unsigned char posPlayer;
-	unsigned short posBoxes[30];
-	unsigned short posGoals[30];
-	unsigned char grid[400];
-	// Quantidade de caixas encontradas
-	unsigned char boxes;
-	// Quantidade de caixas no objetivo
-	unsigned char boxesOnGoals;
-	// Caminho para a solução
-	unsigned char path[MAX_PATH_SIZE];
-	unsigned int pathSize;
-	unsigned int heuristic;
-} State;
 
 // Lista ligada (com hash de resto da divisão) para os ids visitados.
 typedef struct visitedId {
@@ -47,14 +30,8 @@ visitedId *idList[ID_HASH];
 unsigned int activeStates = 0;
 unsigned int storedIds = 0;
 unsigned int filteredIds = 0;
-unsigned int movesBlocked = 0;
 
 Node **last;
-
-void printPath(State *s) {
-	int i = 0;
-	printf("Caminho: %s\n", s->path);
-}
 
 // Função que procura o id na lista
 unsigned char findId(char *idHash, unsigned long long int id) {
@@ -123,7 +100,7 @@ unsigned char getStateId(State *s) {
 		strcat(idHash, buffer);
 		strcat(idHash, " ");
 	}
-	sprintf(buffer, "%d", s->posPlayer);
+	sprintf(buffer, "%hu", s->posPlayer);
 	strcat(idHash, buffer);
 
 	// Procuramos o ID na lista. Se estiver, retornamos verdadeiro
@@ -137,14 +114,6 @@ unsigned char getStateId(State *s) {
 
 	return 0;
 }
-
-void addPath(char c, State *s) {
-	// Se atingimos o limite de tamanho, encerramos
-	if (s->pathSize == MAX_PATH_SIZE) {
-		return;
-	}
-	s->path[s->pathSize++] = c;
-};
 
 int getHeuristic(State *s) {
 	int min, x, y, xGoal, yGoal, dist, h = 0;
@@ -251,11 +220,9 @@ void placeThis(char c, int x, int y, struct State *s) {
 void buildMap(struct State *s, char *level) {
 	s->boxes = 0;
 	s->boxesOnGoals = 0;
-	s->pathSize = 0;
 
 	memset(s->posGoals, 0, 30);
 	memset(s->posBoxes, 0, 30);
-	memset(s->path, 0, 200);
 	memset(idList, 0, ID_HASH);
 
 	last = (Node **)malloc(sizeof(void *));
@@ -298,139 +265,6 @@ void buildMap(struct State *s, char *level) {
 
 	printf("\nO mapa é %d x %d\n", maxWidth, height + 1);
 	printf("O mapa possui %d caixas.\n", s->boxes);
-};
-
-int checkWallsAt(State *s, unsigned int pos) {
-	if (s->grid[pos] == '#') {
-		return 1;
-	}
-	return 0;
-};
-
-int sign(int x) { return (x > 0) - (x < 0); }
-
-unsigned char boxTrapped(State *s, int pos) {
-	int wallsUp = s->grid[pos - 20] == '#',
-	    wallsDown = s->grid[pos + 20] == '#',
-	    wallsLeft = s->grid[pos - 1] == '#',
-	    wallsRight = s->grid[pos + 1] == '#';
-	if (s->grid[pos] == '.') {
-		return 0;
-	}
-	if (wallsUp && wallsLeft) {
-		movesBlocked++;
-		return 1;
-	}
-	if (wallsUp && wallsRight) {
-		movesBlocked++;
-		return 1;
-	}
-	if (wallsDown && wallsLeft) {
-		movesBlocked++;
-		return 1;
-	}
-	if (wallsDown && wallsRight) {
-		movesBlocked++;
-		return 1;
-	}
-
-	return 0;
-}
-
-char movePlayer(struct State *s, int dir) {
-	// dir == 0 -> direita
-	// dir == 1 -> esquerda
-	// dir == 2 -> cima
-	// dir == 3 -> baixo
-	char c;
-	unsigned char box = 0;
-	int tempPos, i = -1;
-	int movingParam;
-	switch (dir) {
-	case direita:
-		movingParam = 1;
-		c = 'r';
-		break;
-	case esquerda:
-		movingParam = -1;
-		c = 'l';
-		break;
-	case cima:
-		movingParam = -20;
-		c = 'u';
-		break;
-	case baixo:
-		movingParam = 20;
-		c = 'd';
-		break;
-	}
-
-	// Usamos o trecho de código para obter o parametro de movimento, assim como
-	// o caracter. Como parâmetro de movimento, queremos transformar a visao 1D
-	// do array no movimento 2D do personagem, ou seja, dado que cada linha
-	// possui 20 colunas, "andar para cima" recua 20 colunas, por isso o
-	// movingParam é -20. A mesma lógica se aplica para os outros
-
-	tempPos = s->posPlayer + movingParam;
-
-	if (checkWallsAt(s, tempPos)) {
-		// Tem uma parede.
-		return 0;
-	}
-
-	if (s->grid[tempPos] == '$' || s->grid[tempPos] == '*') {
-		// Tem uma caixa na direção
-		if (checkWallsAt(s, tempPos + movingParam) == 1 ||
-		    s->grid[tempPos + movingParam] == '$' ||
-		    s->grid[tempPos + movingParam] == '*') {
-			// Tem uma parede ou caixa após a caixa.
-			return 0;
-		};
-		// Deixa a letra maiuscula
-		c -= 32;
-		box = 1;
-	}
-
-	// Efetiva o movimento
-	if (box != 0) {
-		if (boxTrapped(s, tempPos + movingParam)) {
-			return 0;
-		}
-		// Empurrou uma caixa
-		if (s->grid[tempPos] == '*') {
-			// A caixa estava sobre um alvo
-			s->grid[tempPos] = '.';
-			s->boxesOnGoals--;
-		} else {
-			// Não estava sobre nada
-			s->grid[tempPos] = 32;
-		}
-		if (s->grid[tempPos + movingParam] == '.') {
-			// A posição de destino da caixa é alvo
-			s->grid[tempPos + movingParam] = '*';
-			s->boxesOnGoals++;
-		} else {
-			s->grid[tempPos + movingParam] = '$';
-		}
-		// Move a caixa nas posições
-		for (int i = 0; i < 30 && s->posBoxes[i] != 0; i++) {
-			if (s->posBoxes[i] == tempPos) {
-				s->posBoxes[i] += movingParam;
-			}
-		}
-	}
-
-	s->posPlayer = tempPos;
-
-	// Verifica se o Id é único e claro, qual seu valor.
-	if (getStateId(s)) {
-		return 0;
-	}
-
-	addPath(c, s);
-
-	// Retorna o efeito
-	return c;
 };
 
 unsigned char isFinal(State *s) {
@@ -498,7 +332,7 @@ void popState(Node **s, State **rootState) {
 int main(int argc, char *argv[]) {
 	Node *root = (Node *)malloc(sizeof(Node));
 	State *s, *rootState;
-	unsigned char final;
+	unsigned char final = 0;
 
 	root->state = (State *)malloc(sizeof(State));
 	root->nextState = NULL;
@@ -516,8 +350,7 @@ int main(int argc, char *argv[]) {
 			       numberOfNodes, activeStates,
 			       activeStates * (sizeof(Node) + sizeof(State)) / 1048576,
 			       storedIds, storedIds * sizeof(visitedId) / 1048576);
-			printf("%d ids filtrados\n", filteredIds);
-			printf("%d moveBlocks\n\n", movesBlocked);
+			printf("%d ids filtrados\n\n", filteredIds);
 		}
 
 		free(rootState);
@@ -526,7 +359,7 @@ int main(int argc, char *argv[]) {
 
 		for (int i = 0; i < 4; i++) {
 			copyState(rootState, s);
-			if (movePlayer(s, i) != 0) {
+			if (movePlayer(s, i, getStateId) != 0) {
 				final = insertState(&root, s);
 			}
